Minimal Changes still building out diffusion
I'm still struggling with figuring out how to code the diffusion solution.  I know for 1 dimension and based on the results of <x^2> we should have a D value of 1/4 for one dimension as the slope of the <x^2> graph is 1/2.  I am assuming it should be 1/8 for 2-D.  But I can't seem to figure out how to implement the diffusion algorithm from the notes.